* Updated Copyright notices on all files.  * Clean-up.  * src/progname.c: added function getprogname(), renamed set_program_name to setprogname.  * src/syntax.c (syntax): now wraps vwarnx instead of using warnx to print the error message.  * src/error.c: removed.  * build-aux/ar-lib, build-aux/depcomp, build-aux/compile, build-aux/missing, build-aux/install-sh: updated automake/autoconf auxilary files.  * configure.ac: removed tests for features no longer used.
git-svn-id: https://htpmt.googlecode.com/svn/trunk@104 25275c40-b298-11de-9c04-c7668166297d

+++ src/readpasswd.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2008, 2009, 2010 David Winter
+ * Copyright 2008, 2009, 2010, 2011, 2012 David Winter
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -120,7 +120,7 @@ readpasswd(const char *prompt, char *buffer, size_t bufsize, int flags)
 	}
 
 	if (len >= bufsize)
-		errx(EXIT_OVERFLOW, "the password must not contain more than %i characters", bufsize-1);
+		errx(EXIT_OVERFLOW, "the password must not contain more than %u characters", (unsigned int) bufsize-1);
 
 	return buffer;
 }
